Move parsing logic to Parser instance, keep class-level API.
